test(interactive): keep submit text coverage lint-clean
