+++ src/plugins/interactive.test.ts
@@ -946,42 +946,6 @@ describe("plugin interactive handlers", () => {
     expect(afterInvoke).toHaveBeenLastCalledWith({ handled: true });
   });
 
-  it("exposes submitText to channel post-handler processing exactly once", async () => {
-    const result = { handled: true, submitText: "Continue in this thread" };
-    const handler = vi.fn(async () => result);
-    const afterInvoke = vi.fn(async () => {});
-    expect(
-      registerPluginInteractiveHandler("quick-replies-plugin", {
-        channel: "discord",
-        namespace: "quick-replies",
-        handler,
-      }),
-    ).toEqual({ ok: true });
-
-    const baseParams = {
-      ...createDiscordDispatchParams({
-        data: "quick-replies:continue",
-        interactionId: "discord-submit-text",
-      }),
-      afterInvoke,
-    };
-
-    await expect(dispatchInteractive(baseParams)).resolves.toEqual({
-      matched: true,
-      handled: true,
-      duplicate: false,
-      result,
-    });
-    await expect(dispatchInteractive(baseParams)).resolves.toEqual({
-      matched: true,
-      handled: true,
-      duplicate: true,
-    });
-    expect(handler).toHaveBeenCalledTimes(1);
-    expect(afterInvoke).toHaveBeenCalledOnce();
-    expect(afterInvoke).toHaveBeenCalledWith(result);
-  });
-
   it("dedupes concurrent interactive dispatches while a handler is still running", async () => {
     let releaseHandler: (() => void) | undefined;
     const handlerGate = new Promise<void>((resolve) => {
